created first basic plugin
Called cwe1, it handles the case there is an except block and only a
pass statement used. It simply checks if there is only one line in each
except handler, and if so, it will check if it's a pass line. If this is
the case it will display the pylint warning.