implement a rudimentary input sanitzation plugin
Detect if a node is an input() statement. If so then further check if
the variable it's set to is ever used on the LHS of any other statement.
If it isn't then notify the user. In addition, f an Expr or Return statment
contain an input() statement in them also notify.